pyproject.toml
[project]
name = "toxic-classificator"
version = "0.1.0"
description = "Toxic comment classification using Qwen2.5-1.5B-Instruct with LoRA fine-tuning"
authors = [
    {name = "Сыров Артём Андреевич", email = "syrov@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "peft>=0.10.0",
    "datasets>=2.18.0",
    "accelerate>=0.27.0",
    "lightning>=2.2.0",
    "bitsandbytes>=0.43.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.4.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "mlflow>=2.10.0",
    "dvc>=3.40.0",
    "dvc-gdrive>=3.0.0",
    "fire>=0.6.0",
    "onnx>=1.15.0",
    "onnxruntime>=1.17.0",
    "tritonclient[all]>=2.40.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "flake8>=7.0.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
